+++ Runtime/Scripts/VertexBufferConfig.cs
@@ -56,7 +56,7 @@ namespace GLTFast
             }
         }
 
-        public VertexBufferConfig(Report report) : base(report) {}
+        public VertexBufferConfig(ILogger logger) : base(logger) {}
 
         public override unsafe JobHandle? ScheduleVertexJobs(
             VertexInputData posInput,
@@ -93,10 +93,10 @@ namespace GLTFast
                 jobCount += uvInputs.Length;
                 switch (uvInputs.Length) {
                     case 1:
-                        texCoords = new VertexBufferTexCoords<VTexCoord1>(report);
+                        texCoords = new VertexBufferTexCoords<VTexCoord1>(logger);
                         break;
                     default:
-                        texCoords = new VertexBufferTexCoords<VTexCoord2>(report);
+                        texCoords = new VertexBufferTexCoords<VTexCoord2>(logger);
                         break;
                 }
             }
@@ -110,7 +110,7 @@ namespace GLTFast
             hasBones = weightsInput.HasValue && jointsInput.HasValue;
             if(hasBones) {
                 jobCount+=2;
-                bones = new VertexBufferBones(report);
+                bones = new VertexBufferBones(logger);
             }
 
             NativeArray<JobHandle> handles = new NativeArray<JobHandle>(jobCount, defaultAllocator);